requirements.txt
aiogram==3.7
python-dotenv==1.0.0